Clean up glyphs
